+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Highlight and don't back up non-existing bind-mount volumes.
 - Render commands more copy-paste-friendly (w/o borders and w/o hard linebreaks).
+- Support new `docker compose ps --json` format.
 
 ## [2.2.1] -- 2022-11-27
 ### Fixed